requirements.txt
pika==1.1.0
requests==2.23.0
flask==1.1.1
flask-api==2.0
viaa-chassis==0.1.3
lxml==4.5.0
uWSGI==2.0.18